refactor: extract analysis proposal factory
